+++ demo/js/htmlOutput/subDirA/subDirAA/testAA1.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link rel="stylesheet" href="../../../src/resources/output/styles.css">
+    </head>
+    <body>
+<h1 id="filetestaa1">FILE TEST AA1</h1>
+<h1 id="sayinghelloworld">Saying hello world!</h1>
+<p>Our project rocks and needs to output Hello world!</p>
+<h2 id="example">Example</h2>
+<p>When the project function is called then we see the output <span class="successful-test">Hello World</span>.</p>
+<h2 id="example-1">Example</h2>
+<p>When the project function is called then we see the output <span class="failed-test"><span class="failed-test-expected-result">Hella World</span> <span class="failed-test-actual-result">Hello World</span></span>.</p>
+<h2 id="addinganumber">Adding a number</h2>
+<p>Adding a single number is simple.
+Assuming we have a number 5, then if the system adds one, 
+we see the output <span class="successful-test">6</span>.</p>
+<h2 id="addingabook">Adding a book</h2>
+<p>Assuming we have a book with an ISBN number12345, a title Test And Testing and a description A book about tests,
+when we add it to the library, then we get the highest ISBN <span class="successful-test">12345</span></p>
+    </body>
+</html>
