+++ demo/js/htmlOutput/test1.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link rel="stylesheet" href="../../../src/resources/output/styles.css">
+    </head>
+    <body>
+<h1 id="sayinghelloworld">Saying hello world!</h1>
+<p>Our project rocks and needs to output Hello world!</p>
+<h2 id="example">Example</h2>
+<p>When the project function is called then we see the output <span class="successful-test">Hello World</span>.</p>
+    </body>
+</html>
