+++ src/filesearch/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+exports.MarkdownFileLocator = require('./MarkdownFileLocator');
