+++ src/index.js
@@ -1,4 +1,5 @@
 'use strict';
 
+exports.filesearch = require('./filesearch');
 exports.interpreter = require('./interpreter');
 exports.testSuite = require('./testsuite');
