+++ src/interpreter/index.js
@@ -1,4 +1,3 @@
 'use strict';
 
-exports.MarkdownFileLocator = require('./MarkdownFileLocator');
 exports.MarkdownParser = require('./MarkdownParser');
